% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute.ALD.R
\name{compute.ALD}
\alias{compute.ALD}
\title{Compute ALD.}
\usage{
compute.ALD(dat, tolerance = 0.005)
}
\arguments{
\item{dat}{A data.frame with 5 required variables:
\tabular{ll}{
     \code{haplo.freq} \tab A numeric vector of haplotype frequencies.\cr
      \code{locus1} \tab A character vector indentifying the first locus.\cr
      \code{locus2} \tab A character vector indentifying the second locus.\cr
      \code{allele1} \tab A character vector indentifying the allele at locus 1.\cr
      \code{allele2} \tab A character vector indentifying the allele at locus 1.\cr
      }}

\item{tolerance}{A threshold for the sum of the haplotype frequencies.
   If the sum of the haplotype frequencies is greater than 1+tolerance or less
   than 1-tolerance an error is returned.}
}
\value{
The return value is a dataframe with the following components:
 \tabular{ll}{
 \code{locus1}	\tab The name of the first locus.\cr
 \code{locus2}	\tab The name of the second locus.\cr
 \code{F.1}	\tab Homozygosity (expected under HWP) for locus 1.\cr
 \code{F.1.2}	\tab Conditional homozygosity* for locus1 given locus2.\cr
 \code{F.2}	\tab Homozygosity (expected under HWP) for locus 1.\cr
 \code{F.2.1}	\tab Conditional homozygosity* for locus2 given locus1.\cr
 \code{ALD.1.2} \tab Asymmetric LD for locus1 given locus2.\cr
 \code{ALD.2.1} \tab  Asymmetric LD for locus2 given locus1.\cr
 }
 *Overall weighted haplotype-specific homozygosity for the first locus given the second locus.
}
\description{
A function to compute asymmetric Linkage Disequilibrium measures (ALD) for polymorphic genetic data. These measures are identical to the correlation measure (r) for bi-allelic data.
}
\section{Details}{

A warning message is given if the sum of the haplotype frequencies is greater than 1.01 or less than 0.99. The haplotype frequencies that are passed to the function are normalized within the function to sum to 1.0 by dividing each frequency by the sum of the passed frequencies.
}
\examples{
data(hla.freqs)
hla.a_b <- hla.freqs[hla.freqs$locus1=="A" & hla.freqs$locus2=="B",]
compute.ALD(hla.a_b)
hla.freqs$locus <- paste(hla.freqs$locus1, hla.freqs$locus2, sep="-")
compute.ALD(hla.freqs[hla.freqs$locus=="C-B",])
}

